replaced int type for size t

+++ 0x1E-search_algorithms/1-binary.c
@@ -8,9 +8,9 @@
  */
 int binary_search(int *array, size_t size, int value)
 {
-	int middle;
-	int start = 0;
-	int end = size - 1;
+	size_t middle;
+	size_t start = 0;
+	size_t end = size - 1;
 
 	if (array == NULL)
 		return (-1);
